Split out and align environment variables between consoles
