+++ js/apps/account-ui/src/root/Root.tsx
@@ -1,10 +1,9 @@
-import {
-  KeycloakProvider,
-  environmentAccount as environment,
-} from "@keycloak/keycloak-ui-shared";
+import { KeycloakProvider } from "@keycloak/keycloak-ui-shared";
 import { Page, Spinner } from "@patternfly/react-core";
 import { Suspense } from "react";
 import { Outlet } from "react-router-dom";
+
+import { environment } from "../environment";
 import { Header } from "./Header";
 import { PageNav } from "./PageNav";
