+++ js/apps/admin-ui/pom.xml
@@ -124,15 +124,15 @@
 <![CDATA[
   <script id="environment" type="application/json">
     {
+      "authServerUrl": "${authServerUrl}",
       "realm": "${loginRealm!"master"}",
       "clientId": "${clientId}",
-      "authServerUrl": "${authServerUrl}",
-      "consoleBaseUrl": "${consoleBaseUrl}",
       "resourceUrl": "${resourceUrl}",
-      "masterRealm": "${masterRealm}",
-      "resourceVersion": "${resourceVersion}",
       "logo": "${properties.logo!""}",
-      "logoUrl": "${properties.logoUrl!""}"
+      "logoUrl": "${properties.logoUrl!""}",
+      "consoleBaseUrl": "${consoleBaseUrl}",
+      "masterRealm": "${masterRealm}",
+      "resourceVersion": "${resourceVersion}"
     }
   </script>
 </body>
