+++ js/apps/admin-ui/src/admin-client.ts
@@ -3,8 +3,9 @@ import {
   createNamedContext,
   useRequiredContext,
 } from "@keycloak/keycloak-ui-shared";
-import { AdminEnvironment } from "@keycloak/keycloak-ui-shared/dist/context/environment";
 import type Keycloak from "keycloak-js";
+
+import type { Environment } from "./environment";
 
 export type AdminClientProps = {
   keycloak: Keycloak;
@@ -19,7 +20,7 @@ export const useAdminClient = () => useRequiredContext(AdminClientContext);
 
 export async function initAdminClient(
   keycloak: Keycloak,
-  environment: AdminEnvironment,
+  environment: Environment,
 ) {
   const adminClient = new KeycloakAdminClient();
