+++ js/libs/ui-shared/src/main.ts
@@ -7,11 +7,8 @@ export {
   type KeycloakContext,
 } from "./context/KeycloakContext";
 export {
-  environmentAccount,
-  environmentAdmin,
-  type AccountEnvironment,
-  type AdminEnvironment,
-  type Feature,
+  getInjectedEnvironment,
+  type BaseEnvironment,
 } from "./context/environment";
 export { ContinueCancelModal } from "./continue-cancel/ContinueCancelModal";
 export {
